# docs/api/swagger.json
{
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This document describes the RESTful APIs provided or assumed by the\nKubernetes Application Runtime (KAR). It consists of two sets of APIs:\n+ The **sidecar** APIs are implemented by the KAR sidecar.\n+ The **application component** APIs are meant to be implemented by application components.\n\nThe **sidecar** APIs are divided into logical groups:\n+ **Actors**: APIs to invoke actor methods.\n+ **Services**: APIs to invoke service endpoints.\n+ **Callbacks**: APIs to await the response to an asynchronous actor or service invocation.\n+ **Events**: APIs to publish to event sinks or subscribe actors to event sources.\n+ **Reminders**: APIs to schedule future actor invocations.\n+ **State**: APIs to manage the persistent state of actors.\n+ **System**: APIs for controlling the KAR runtime mesh.\n\nThe **application component** APIs are divided into logical groups:\n+ **Actor runtime**: APIs invoked by the sidecar to manage actor instances\nhosted by the application component.",
    "title": "KAR",
    "version": "v1"
  },
  "basePath": "/kar",
  "paths": {
...
    "/v1/actor/{actorType}/{actorId}": {
      "delete": {
        "description": "### Completely remove an actor instance\n\nAll user-level and runtime state of the actor instance indicated by\n`actorType` and `actorId` will be asynchronously deleted.",
        "schemes": [
          "http"
        ],
        "tags": [
          "actors"
        ],
        "summary": "actor",
        "operationId": "idActorDelete",
        "responses": {
          "202": {
            "$ref": "#/responses/response202"
          },
          "404": {
            "$ref": "#/responses/response404"
          },
          "500": {
            "$ref": "#/responses/response500"
          }
        }
      }
    },
    "/v1/actor/{actorType}/{actorId}/call/{methodName}": {
      "post": {
        "description": "### Invoke an actor method\n\nCall invokes the `methodName` method of the\nactor instance indicated by `actorType` and `actorId`.\nThe request body must be a (possibly zero-length) JSON array whose elements\nare used as the actual parameters of the actor method.\nThe result of the call is the result of invoking the target actor method\nunless the `async` or `promise` pragma header is specified.  If the actor\nmethod returns `void` or `undefined`, then a 204 - No Content reponse is returned.",
        "consumes": [
          "application/kar+json"
        ],
        "produces": [
          "application/kar+json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "actors"
        ],
        "summary": "call",
        "operationId": "idActorCall",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ActorType",
            "description": "The actor type",
            "name": "actorType",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ActorID",
            "description": "The actor instance id",
            "name": "actorId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Optionally specify the `async` pragma to make a non-blocking call.\nOptionally specify the `promise` pragma to make a non-blocking call and\nobtain a request id to query later.",
            "name": "Pragma",
            "in": "header"
          },
          {
            "type": "string",
            "example": "computeMyResult",
            "x-go-name": "MethodName",
            "description": "The actor method to be invoked",
            "name": "methodName",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Session",
            "description": "Optionally specific the session to use when performing the call.  Enables re-entrancy for nested actor calls and tracking parent-child relationships",
            "name": "session",
            "in": "query"
          },
          {
            "example": "[3, 'hello', { msg: 'Greetings' }]",
            "description": "A possibly empty array containing the arguments with which to invoke the target actor method.",
            "name": "ActorMethodArguments",
            "in": "body",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/response200CallActorResult"
          },
          "202": {
            "$ref": "#/responses/response202"
          },
          "204": {
            "$ref": "#/responses/response204ActorNoContentResult"
          },
          "404": {
            "$ref": "#/responses/response404"
          },
          "500": {
            "$ref": "#/responses/response500"
          },
          "503": {
            "$ref": "#/responses/response503"
          }
        }
      }
    },
    "/v1/actor/{actorType}/{actorId}/events": {
      "get": {
        "description": "### Get all subscriptions\n\nThis operation returns all subscriptions for the actor instance specified in the path.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "events"
        ],
        "summary": "subscriptions",
        "operationId": "idActorSubscriptionGetAll",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ActorType",
            "description": "The actor type",
            "name": "actorType",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ActorID",
            "description": "The actor instance id",
            "name": "actorId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/response200SubscriptionGetAllResult"
          },
          "500": {
            "$ref": "#/responses/response500"
          },
          "503": {
            "$ref": "#/responses/response503"
          }
        }
      },
      "delete": {
        "description": "### Cancel all subscriptions\n\nThis operation cancels all subscriptions for the actor instance specified in the path.\nThe number of subscriptions cancelled is returned as the result of the operation.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "events"
        ],
        "summary": "subscriptions",
        "operationId": "idActorSubscriptionCancelAll",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ActorType",
            "description": "The actor type",
            "name": "actorType",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ActorID",
            "description": "The actor instance id",
            "name": "actorId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/response200SubscriptionCancelAllResult"
          },
          "500": {
            "$ref": "#/responses/response500"
          },
          "503": {
            "$ref": "#/responses/response503"
          }
        }
      }
    },
    "/v1/actor/{actorType}/{actorId}/events/{subscriptionId}": {
      "get": {
        "description": "### Get a subscription\n\nThis operation returns the subscription for the actor instance specified in the path.\nIf there is no subscription with the id `subscriptionId` a `404` response will be returned\nunless the boolean query parameter `nilOnAbsent` is set to `true`.\nIf `nilOnAbsent` is true the `404` response will be replaced with\na `200` response with a `nil` response body.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "events"
        ],
        "summary": "subscriptions/id",
        "operationId": "idActorSubscriptionGet",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ActorType",
            "description": "The actor type",
            "name": "actorType",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ActorID",
            "description": "The actor instance id",
            "name": "actorId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "SubscriptionID",
            "description": "The id of the specific subscription being targeted",
            "name": "subscriptionID",
            "in": "path",
            "required": true
          },
          {
            "type": "boolean",
            "x-go-name": "ErrorOnAbsent",
            "description": "Replace a REST-style `404` response with a `200` and nil response body when the requested key is not found.",
            "name": "nilOnAbsent",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/response200SubscriptionGetResult"
          },
          "404": {
            "$ref": "#/responses/response404"
          },
          "500": {
            "$ref": "#/responses/response500"
          },
          "503": {
            "$ref": "#/responses/response503"
          }
        }
      },
      "put": {
        "description": "### Subscribe to a topic\n\nSubscribe the actor instance using the subscriptionId specified in the path\nas described by the data provided in the request body.\nIf there is already a subscription for the target actor instance with the same subscriptionId,\nthat existing subscription will be updated based on the request body.\nThe operation will not return until after the actor instance is subscribed.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "events"
        ],
        "summary": "subscriptions/id",
        "operationId": "idActorSubscribe",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ActorType",
            "description": "The actor type",
            "name": "actorType",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ActorID",
            "description": "The actor instance id",
            "name": "actorId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "SubscriptionID",
            "description": "The id of the specific subscription being targeted",
            "name": "subscriptionID",
            "in": "path",
            "required": true
          },
          {
            "description": "The request body describes the subscription",
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/EventSubscribeOptions"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/response201"
          },
          "204": {
            "$ref": "#/responses/response204"
          },
          "500": {
            "$ref": "#/responses/response500"
          },
          "503": {
            "$ref": "#/responses/response503"
          }
        }
      },
      "delete": {
        "description": "### Cancel a subscription\n\nThis operation cancels the subscription for the actor instance specified in the path.\nIf the subscription is successfully cancelled a `200` response with a body of `1` will be returned.\nIf the subscription is not found, a `404` response will be returned unless\nthe boolean query parameter `nilOnAbsent` is set to `true`. If `nilOnAbsent`\nis sent to true the `404` response will instead be a `200` with a body containing `0`.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "events"
        ],
        "summary": "subscriptions/id",
        "operationId": "idActorSubscriptionCancel",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ActorType",
            "description": "The actor type",
            "name": "actorType",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ActorID",
            "description": "The actor instance id",
            "name": "actorId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "SubscriptionID",
            "description": "The id of the specific subscription being targeted",
            "name": "subscriptionID",
            "in": "path",
            "required": true
          },
          {
            "type": "boolean",
            "x-go-name": "ErrorOnAbsent",
            "description": "Replace a REST-style `404` response with a `200` and nil response body when the requested key is not found.",
            "name": "nilOnAbsent",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/response200SubscriptionCancelResult"
          },
          "404": {
            "$ref": "#/responses/response404"
          },
          "500": {
            "$ref": "#/responses/response500"
          },
          "503": {
            "$ref": "#/responses/response503"
          }
        }
      }
    },
...
    "/v1/event/{topic}": {
      "put": {
        "description": "### Creates or updates a given topic\n\nParameters are specified in the body of the post, as stringified JSON.\nNo body passed causes a default creation.",
        "consumes": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "events"
        ],
        "summary": "topic",
        "operationId": "idTopicCreate",
        "parameters": [
          {
            "description": "The request body describes the topic to be created",
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/topicCreateOptions"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/response201"
          },
          "204": {
            "$ref": "#/responses/response204"
          },
          "500": {
            "$ref": "#/responses/response500"
          }
        }
      },
      "delete": {
        "description": "### Deletes given topic\n\nDeletes topic specified in route.",
        "consumes": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "events"
        ],
        "summary": "topic",
        "operationId": "idTopicDelete",
        "responses": {
          "200": {
            "$ref": "#/responses/response200"
          },
          "500": {
            "$ref": "#/responses/response500"
          }
        }
      }
    },
...
  },
...
  "responses": {
...
    "response200CallActorResult": {
      "description": "The result of invoking the actor method",
      "schema": {
        "$ref": "#/definitions/actorCallResult"
      },
      "headers": {
        "Body": {
          "description": "The result returned by the actor method"
        }
      }
    },
...
    "response200SubscriptionGetResult": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/source"
      },
      "headers": {
        "Body": {
          "description": "The subscription"
        }
      }
    },
...
    "response202": {
      "description": "An asynchronous operation has been accepted for eventual execution"
    },
...
    "response204ActorNoContentResult": {
      "description": "Indicates that the result of an actor method is void or undefined."
    },
...
    "response404": {
      "description": "Response indicating requested resource is not found",
      "headers": {
        "body": {
          "type": "string",
          "example": "Not Found",
          "description": "Requested resource is not found"
        }
      }
    },
    "response500": {
      "description": "A message describing the error",
      "headers": {
        "body": {
          "type": "string",
          "example": "Internal Server Error",
          "description": "A message describing the error"
        }
      }
    },
    "response503": {
      "description": "A message describing the error",
      "headers": {
        "body": {
          "type": "string",
          "example": "Service Unavailable",
          "description": "A message describing the error"
        }
      }
    },
...
  },
...
}